Remove helper functions file
